requirements.txt
langchain~=0.1.20
langchain-groq
python-dotenv~=1.0.0
unstructured~=0.14.5
google-api-python-client==2.132.0
google-auth-oauthlib==1.2.0
faiss-cpu
